Relation

+++ xp/src/main/java/xp/nova/sqldb/graph/Relation_.java
@@ -2,5 +2,5 @@ package xp.nova.sqldb.graph;
 
 public interface Relation_
 {
-    public int getValue();
+    public long getValue();
 }
